♻️ Refactor: 스타일 파일들을 base와 components 폴더로 구조화
